@@ -73,6 +73,7 @@
     "react-simple-code-editor": "^0.9.10",
     "react-sortable-hoc": "^1.10.1",
     "reactcss": "^1.2.3",
+    "react-dom": "^18.2.0",
     "robust-point-in-polygon": "^1.0.3",
     "slugid": "^3.2.0",
     "url-parse": "^1.4.3",
@@ -105,7 +106,7 @@
     "@types/pako": "^2.0.2",
     "@types/prop-types": "^15.7.3",
     "@types/react": "^17.0.2",
-    "@types/react-dom": "^17.0.2",
+    "@types/react-dom": "^18.2.0",
     "@vitejs/plugin-react": "^4.2.0",
     "@web/test-runner": "^0.17.2",
     "@web/test-runner-puppeteer": "^0.14.1",
@@ -123,8 +124,8 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "pixi.js": "^6.5.2",
     "prettier": "^2.5.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "sass": "^1.54.5",
     "tinyspy": "^1.0.2",
     "typescript": "^5.2.2",
